% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{age_df}
\alias{age_df}
\title{Age Trajectory Data}
\format{
\code{"age_df"} is a simulated data frame with 712 cases (rows) and 56 variables
(columns). 51 rois are included.
}
\usage{
data(age_df)
}
\description{
This data set is used to observe life span age trend of brain structures.
}
\examples{
data(age_df)
head(age_df)
}
\keyword{datasets}
