% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adni}
\alias{adni}
\title{Harmonization Data}
\format{
\code{"adni"} is a data frame with 2515 cases (rows) and 104 variables
(columns). 62 features are included for harmonization purpose.
}
\source{
\href{https://adni.loni.usc.edu/}{ADNI}
}
\usage{
data(adni)
}
\description{
This data set is a sample data set from ADNI study.
}
\examples{
data(adni)
head(adni)
}
\keyword{datasets}
