% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Help_Func.R
\name{age_save}
\alias{age_save}
\title{Export Brain ROI Age Trends}
\usage{
age_save(path, age_list)
}
\arguments{
\item{path}{The path to save the excel file.}

\item{age_list}{A list containing all ROIs' true volumes, age trend estimates, and the fitted GAMLSS model.}
}
\value{
This function does not return a value. It saves the data to the specified file.
}
\description{
Save all brain age trends into a single Excel file.
}
\examples{
# Initialize result to NULL for safety
age_list <- NULL

# Check if the previous results file exists and load it
if (file.exists("./tests/testthat/previous-results/age_list.rds")) {
  age_list <- readRDS("./tests/testthat/previous-results/age_list.rds")
}

# Use the result if it is available
if (!is.null(age_list)) {
  temp_dir <- tempdir()
  age_save(temp_dir, age_list)
  message("Age trend table saved to: ", temp_dir)
} else {
  message("Age list is NULL. Please ensure the file exists and is accessible.")
}

}
