% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_shiny.R
\name{age_shiny}
\alias{age_shiny}
\title{Lifespan Age Trends}
\usage{
age_shiny(age_list, features, quantile_type)
}
\arguments{
\item{age_list}{A list contains all rois' true volumes and age trend estimates.}

\item{features}{A vector of roi names.}

\item{quantile_type}{A vector of quantile types.}
}
\description{
Provide estimated lifespan age trends of neuroimaging-derived brain structures through shiny app.
}
