% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comfam_shiny.R
\name{comfam_shiny}
\alias{comfam_shiny}
\title{Batch Effect Interactive Visualization}
\usage{
comfam_shiny(result, after = FALSE, ...)
}
\arguments{
\item{result}{A list derived from \code{visual_prep()} that contains datasets for shiny visualization.}

\item{after}{A boolean variable indicating whether the dataset is before or after harmonization. The default value is FALSE.}

\item{...}{Additional arguments to \code{comfam} or \code{covfam} models.}
}
\description{
Provide interactive batch/site effect visualization through shiny app.
}
