% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comfam_shiny.R
\name{comfam_shiny}
\alias{comfam_shiny}
\title{Batch Effect Interactive Visualization}
\usage{
comfam_shiny(result)
}
\arguments{
\item{result}{A list derived from \code{visual_prep()} that contains datasets for shiny visualization.}
}
\description{
Provide interactive batch/site effect visualization through shiny app.
}
