% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_shiny.R
\name{cus_result_gen}
\alias{cus_result_gen}
\title{Generate Customized Predicted Quantiles List}
\usage{
cus_result_gen(age_list, customized_q = 0.75, f)
}
\arguments{
\item{age_list}{A list containing all ROIs' true volumes, age trend estimates, and the fitted GAMLSS model.}

\item{customized_q}{A numeric value between 0 and 1 representing the quantile to predict (e.g., \code{0.75} for the 75th percentile).}

\item{f}{A string specifying the feature of interest within the \code{age_list}.}
}
\value{
A list containing the customized quantile value used for predictions and a data frame with columns for age, quantile type, prediction, and sex.
}
\description{
This function computes customized predicted quantiles for a specified feature across both female and male groups
using a GAMLSS model. The resulting list object is structured for visualization purposes.
}
\details{
This function utilizes the GAMLSS model to compute predictions for the specified quantile across demographic groups.
The results include predictions for both the specified quantile and the median, enabling enhanced visualization.
}
\examples{
# Initialize result to NULL for safety
age_list <- NULL

# Check if the previous results file exists and load it
if (file.exists("./tests/testthat/previous-results/age_list.rds")) {
  age_list <- readRDS("./tests/testthat/previous-results/age_list.rds")
}

# Use the result if it is available
if (!is.null(age_list)) {
  cus_result_gen(age_list, customized_q = 0.75, f = "Volume_1")
} else {
  message("Age list is NULL. Please ensure the file exists and is accessible.")
}
}
