% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_shiny.R
\name{customize_percentile}
\alias{customize_percentile}
\title{Generate Predicted Quantiles for Age Trends}
\usage{
customize_percentile(age_list, feature, q = 0.75, s = "F")
}
\arguments{
\item{age_list}{A list containing all ROIs' true volumes, age trend estimates, and the fitted GAMLSS model.}

\item{feature}{A string specifying the feature of interest within the \code{age_list}.}

\item{q}{A numeric value between 0 and 1 representing the quantile to predict (e.g., \code{0.5} for the median).}

\item{s}{A string indicating the gender of the group for which the predictions are generated (e.g., \code{"F"} for female, \code{"M"} for male).}
}
\value{
A data frame containing columns for age, quantile type, prediction, and sex.
}
\description{
This function computes predicted quantiles for a specified feature and demographic group based on a GAMLSS model.
The function interpolates predictions over a range of ages while accounting for fixed covariates.
}
\details{
This function uses a GAMLSS model to generate predictions for a specified quantile and demographic group.
The predictions are computed over a sequence of ages (\code{age_test}) that spans the observed age range in the data.
The function adjusts for fixed covariates such as \code{icv} by using their mean values from the input data.
}
\examples{
# Initialize result to NULL for safety
age_list <- NULL

# Check if the previous results file exists and load it
if (file.exists("./tests/testthat/previous-results/age_list.rds")) {
  age_list <- readRDS("./tests/testthat/previous-results/age_list.rds")
}

# Use the result if it is available
if (!is.null(age_list)) {
  customize_percentile(age_list, feature = "Volume_1", q = 0.5, s = "F")
} else {
  message("Age list is NULL. Please ensure the file exists and is accessible.")
}
}
