% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Help_Func.R
\name{diag_save}
\alias{diag_save}
\title{Export Batch Effect Diagnosis Results}
\usage{
diag_save(path, result)
}
\arguments{
\item{path}{The path to save the excel file.}

\item{result}{A list derived from \code{visual_prep()} that contains datasets and statistical test results.}
}
\value{
This function does not return a value. It saves the data to the specified file.
}
\description{
Save all the batch effect diagnosis result in a single excel file.
}
\examples{
\dontrun{
result <- readRDS("./tests/testthat/previous-results/lm_result.rds")
temp_dir <- tempdir()
diag_save(temp_dir, result)
}
}
