% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual_prep.R
\name{model_gen}
\alias{model_gen}
\title{Model generations}
\usage{
model_gen(
  y,
  type,
  batch,
  covariates,
  interaction = NULL,
  random = NULL,
  smooth = NULL,
  df
)
}
\arguments{
\item{y}{Dependent variable in the model.}

\item{type}{A model function name that is used or to be used in the ComBatFamily Package (eg: "lmer", "lm").}

\item{batch}{Factor indicating batch (often equivalent to site or scanner).}

\item{covariates}{Name of covariates supplied to \code{model}.}

\item{interaction}{Expression of interaction terms supplied to \code{model} (eg: "age*diagnosis").}

\item{random}{Variable name of a random effect in linear mixed effect model.}

\item{smooth}{Variable name that requires a smooth function.}

\item{df}{Dataset to be harmonized.}
}
\value{
A model object
}
\description{
Generate appropriate regression models based on the model type and formula
}
